Create pcai.py
This code will first get the activity data from the MetaMask API. It will then get the receiving wallet addresses and the sending wallet address from the activity data. Finally, it will check if the receiving wallet has the sending wallet address. If it does, the transaction will be allowed. Otherwise, the transaction will be blocked.